Change directory structure and use Axios for HTTP requests
Instead of the Rails-style file type first layout, the app now uses a
feature oriented structure, but the containers and components remain
separate from the Redux features.

I don't want the app to end up in a place where some Redux files are
inside of container directories and then a bunch of stuff ends up in a
`common/` directory, so instead, all of the Redux files are shared by
default, but organized based on what they do.

Of course, the division of features for this app is contrived for the
sake of illustration. In reality, creating a list of Github usernames
to lookup and looking up the usernames would all be part of one
feature, but they're separate here, just to show how it could be done
in a more complex app.